Cleaned up config integration
