+++ src/main/java/dev/micalobia/full_slabs/FullSlabsModClient.java
@@ -10,9 +10,9 @@ import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
 import net.fabricmc.fabric.api.client.model.ModelLoadingRegistry;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.client.option.KeyBinding;
+import net.minecraft.client.render.RenderLayer;
 import net.minecraft.client.util.InputUtil.Type;
 import org.lwjgl.glfw.GLFW;
-import net.minecraft.client.render.RenderLayer;
 
 public class FullSlabsModClient implements ClientModInitializer {
 	private static KeyBinding toggleWidget;
