+++ src/main/java/dev/micalobia/full_slabs/client/render/model/FullSlabModel.java
@@ -1,18 +1,16 @@
 package dev.micalobia.full_slabs.client.render.model;
 
 import com.mojang.datafixers.util.Pair;
-import dev.micalobia.full_slabs.FullSlabsMod;
 import dev.micalobia.full_slabs.block.FullSlabBlock;
 import dev.micalobia.full_slabs.util.Utility;
+import net.fabricmc.fabric.api.renderer.v1.Renderer;
 import net.fabricmc.fabric.api.renderer.v1.RendererAccess;
 import net.fabricmc.fabric.api.renderer.v1.material.BlendMode;
 import net.fabricmc.fabric.api.renderer.v1.material.MaterialFinder;
 import net.fabricmc.fabric.api.renderer.v1.material.RenderMaterial;
 import net.fabricmc.fabric.api.renderer.v1.mesh.MeshBuilder;
 import net.fabricmc.fabric.api.renderer.v1.mesh.QuadEmitter;
-import net.fabricmc.fabric.api.renderer.v1.model.FabricBakedModel;
 import net.fabricmc.fabric.api.renderer.v1.render.RenderContext;
-import net.fabricmc.fabric.api.renderer.v1.render.RenderContext.QuadTransform;
 import net.fabricmc.fabric.api.rendering.data.v1.RenderAttachedBlockView;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
@@ -20,7 +18,10 @@ import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.render.RenderLayer;
 import net.minecraft.client.render.RenderLayers;
 import net.minecraft.client.render.block.BlockRenderManager;
+import net.minecraft.client.render.model.BakedModel;
+import net.minecraft.client.render.model.BakedQuad;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.Direction;
 import net.minecraft.util.math.Direction.Axis;
 import net.minecraft.world.BlockRenderView;
 
@@ -37,7 +38,7 @@ public class FullSlabModel extends BasicModel {
 		BlockState positiveState = Utility.getSlabState(pair, axis, true);
 		BlockState negativeState = Utility.getSlabState(pair, axis, false);
 		BlockRenderManager manager = MinecraftClient.getInstance().getBlockRenderManager();
-		var renderer = RendererAccess.INSTANCE.getRenderer();
+		Renderer renderer = RendererAccess.INSTANCE.getRenderer();
 		assert renderer != null;
 		MaterialFinder finder = renderer.materialFinder();
 		RenderLayer positiveLayer = RenderLayers.getBlockLayer(positiveState);
@@ -69,10 +70,5 @@ public class FullSlabModel extends BasicModel {
 			quadEmitter.emit();
 		}
 		context.meshConsumer().accept(builder.build());
-//		context.fallbackConsumer().accept((BakedModel) positiveModel);
-//		context.fallbackConsumer().accept((BakedModel) negativeModel);
-		//positiveModel.emitBlockQuads(blockView, state, pos, randomSupplier, context);
-		//negativeModel.emitBlockQuads(blockView, state, pos, randomSupplier, context);
-		//context.popTransform();
 	}
 }
