+++ src/main/java/dev/micalobia/full_slabs/mixin/item/BlockItemMixin.java
@@ -3,7 +3,7 @@ package dev.micalobia.full_slabs.mixin.item;
 import dev.micalobia.full_slabs.FullSlabsMod;
 import dev.micalobia.full_slabs.block.ExtraSlabBlock;
 import dev.micalobia.full_slabs.block.entity.ExtraSlabBlockEntity;
-import dev.micalobia.full_slabs.block.entity.ExtraSlabBlockEntity.SlabExtra;
+import dev.micalobia.full_slabs.config.SlabExtra;
 import dev.micalobia.full_slabs.util.MixinSelf;
 import dev.micalobia.full_slabs.util.Utility;
 import net.minecraft.block.Block;
